requirements.txt
django-allauth==0.43.0
pysaml2==5.0.0
